Project views in separate blueprint
