+++ dribdat/public/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 """The public module, including the homepage, API and user auth."""
-from . import views, auth, api  # noqa
+from . import views, project, auth, api  # noqa
