+++ dribdat/templates/admin/eventprint.html
@@ -10,7 +10,7 @@
     <div class="project-info">
 
       <ul style="float:right; min-width:50%; list-style-type:none; border-left:1px solid #999; padding-left:15px">
-        <li>Permalink: <a href="{{ url_for('public.project', project_id=project.id) }}">{{ url_for('public.project', project_id=project.id, _external=True) }}</a></li>
+        <li>Permalink: <a href="{{ url_for('project.project', project_id=project.id) }}">{{ url_for('project.project', project_id=project.id, _external=True) }}</a></li>
         {% if project.webpage_url %}
         <li>Homepage: <a href="{{ project.webpage_url }}">{{ project.webpage_url }}</a></li>
         {% endif %}
