+++ dribdat/templates/admin/project.html
@@ -9,10 +9,10 @@
 
     <hr>
 
-    <a href="{{ url_for('public.project', project_id=project.id) }}" class="btn btn-sm btn-default">
+    <a href="{{ url_for('project.project', project_id=project.id) }}" class="btn btn-sm btn-default">
       View
     </a>
-    <a href="{{ url_for('public.project_edit', project_id=project.id) }}" class="btn btn-sm btn-default">
+    <a href="{{ url_for('project.project_edit', project_id=project.id) }}" class="btn btn-sm btn-default">
       Edit details
     </a>
