+++ dribdat/templates/admin/projects.html
@@ -42,7 +42,7 @@
               <small><br>{{ project.user.username }}</small></td>
             <td>{{ project.updated_at|since_date }}</td>
             <td><div class="btn-group">
-              <a href="{{ url_for('public.project', project_id=project.id) }}" class="btn btn-sm btn-primary">
+              <a href="{{ url_for('project.project', project_id=project.id) }}" class="btn btn-sm btn-primary">
                 View
               </a>
               {% if project.is_hidden %}
@@ -51,7 +51,7 @@
                 </a>
               {% else %}
                 {% if project.is_autoupdate %}
-                <a href="{{ url_for('public.project_autoupdate', project_id=project.id) }}" class="btn btn-sm btn-warning">
+                <a href="{{ url_for('project.project_autoupdate', project_id=project.id) }}" class="btn btn-sm btn-warning">
                   Sync
                 </a>
                 {% endif %}
