+++ dribdat/templates/public/dribs.html
@@ -21,7 +21,7 @@
         <div class="timeline-img"></div>
         <div class="timeline-content timeline-card js--fadeInBottom">
           <div class="timeline-img-header">
-            <a href="{{ url_for('public.project', project_id=s.project.id)}}">
+            <a href="{{ url_for('project.project', project_id=s.project.id)}}">
               <h2>{{s.project.name}}</h2>
             </a>
           </div>
