+++ dribdat/templates/public/embed.html
@@ -18,7 +18,7 @@
       {% endif %}
     </a>
     {% if current_event.can_start_project %}
-    <a href="{{ url_for('public.project_new', event_id=current_event.id) }}" target="_blank" class="btn btn-success">
+    <a href="{{ url_for('project.project_new', event_id=current_event.id) }}" target="_blank" class="btn btn-success">
       Share project
       <i class="fa fa-rocket"></i>
     </a>
@@ -27,7 +27,7 @@
   <div class="honeycomb">
     <div class="ibws-fix">
     {% for project in projects %}
-      <a href="{{ url_for('public.project', project_id=project.id) }}" target="_blank"
+      <a href="{{ url_for('project.project', project_id=project.id) }}" target="_blank"
         class="hexagon project" category-id="{{project.category_id}}"
         style="background-color:{{project.logo_color}}; background-image:url('{{project.image_url}}')">
         <div class="hexagontent">
