+++ dribdat/templates/public/event.html
@@ -57,7 +57,7 @@
   <a name="projects"></a>
   <div class="ibws-fix">
   {% for project in projects %}
-    <a href="{{ url_for('public.project', project_id=project.id) }}"
+    <a href="{{ url_for('project.project', project_id=project.id) }}"
        title="{{ project.summary }}"
        class="hexagon {{ 'challenge' if project.phase=='Challenge' else 'project' }} category-highlight"
        category-id="{{ project.category_id }}"
@@ -107,7 +107,7 @@
 
 {% if current_event.can_start_project %}
   <div class="start-project btn-group">
-    <a href="{{ url_for('public.project_new', event_id=current_event.id) }}" class="btn btn-lg btn-success">
+    <a href="{{ url_for('project.project_new', event_id=current_event.id) }}" class="btn btn-lg btn-success">
       {% if current_event.has_started %}
         <i class="fa fa-rocket"></i>
         <span>Get started</span>
