+++ dribdat/templates/public/home.html
@@ -54,7 +54,7 @@
           {% endif %}
         </b></a>
         {% if current_event.has_started and current_event.can_start_project %}
-          <!-- <a href="{{ url_for('public.project_new', event_id=current_event.id) }}" class="btn btn-lg btn-success">
+          <!-- <a href="{{ url_for('project.project_new', event_id=current_event.id) }}" class="btn btn-lg btn-success">
             <i class="fa fa-rocket"></i>
             Start Project</a> -->
         {% endif %}
