+++ dribdat/templates/public/projectpost.html
@@ -6,12 +6,12 @@
 {% block body_class %}projectform{% endblock %}
 
 {% block content %}
-  <h2>Drib on
-    <a class="project-link" href="{{ url_for('public.project', project_id=project.id) }}">{{project.name}}</a>
+  <h2>Drib about
+    <a class="project-link" href="{{ url_for('project.project', project_id=project.id) }}">{{project.name}}</a>
   </h2>
   <div class="form-project-post">
 
-    {{ render_form(url_for('public.project_post', project_id=project.id), form) }}
+    {{ render_form(url_for('project.project_post', project_id=project.id), form) }}
 
   </div>
 {% endblock %}
