+++ dribdat/templates/public/userprofile.html
@@ -100,7 +100,7 @@
         <div class="row flex-row flex-nowrap">
         {% for project in projects %}
             <a class="col-5 card project"
-               href="{{ url_for('public.project', project_id=project.id) }}">
+               href="{{ url_for('project.project', project_id=project.id) }}">
 
               {% if project.image_url %}
                 <img class="project-image" src="{{project.image_url}}">
@@ -135,7 +135,7 @@
         <div class="row flex-row flex-nowrap">
         {% for project in challenges %}
             <a class="col-5 card project"
-               href="{{ url_for('public.project', project_id=project.id) }}">
+               href="{{ url_for('project.project', project_id=project.id) }}">
 
               <div class="card-body">
                 <h5 class="card-title">{{ project.name }}</h5>
@@ -168,7 +168,7 @@
     <div class="timeline-img"></div>
     <div class="timeline-content timeline-card js--fadeInBottom">
       <div class="timeline-img-header">
-        <a href="{{ url_for('public.project', project_id=s.project_id) }}">
+        <a href="{{ url_for('project.project', project_id=s.project_id) }}">
           <h2>{{s.project_name}}</h2>
         </a>
       </div>
